refactor: move the config file path finding logic to config.js
